app.js 깔끔하게 정리
